+++ plugins/ocr/__init__.py
@@ -1,0 +1,6 @@
+# plugins/ocr/__init__.py - v3.0.3
+# OCR Plugin - Image to Text Recognition
+
+from .plugin import Plugin
+
+__all__ = ['Plugin']